+++ packages/kakao_flutter_sdk_auth/lib/src/utils.dart
@@ -7,12 +7,3 @@ String generateRandomString(int length) {
   return String.fromCharCodes(
       Iterable.generate(length, (_) => ch.codeUnitAt(r.nextInt(ch.length))));
 }
-
-/// @nodoc
-extension ListParameterExtension<T> on List<T>? {
-  String? joinToString([String separator = ""]) {
-    if (this == null || this!.isEmpty) return null;
-
-    return this!.join(separator);
-  }
-}