+++ packages/kakao_flutter_sdk_common/lib/src/web/user.dart
@@ -1,6 +1,6 @@
 import 'package:kakao_flutter_sdk_common/src/kakao_sdk.dart';
 
-String createSelectShippingAddressesUrl(final Map<String, dynamic> params) {
+String createselectShippingAddressUrl(final Map<String, dynamic> params) {
   return Uri.https(KakaoSdk.hosts.apps, '/user/address', params).toString();
 }