+++ packages/kakao_flutter_sdk_friend/lib/src/model/selected_user.g.dart
@@ -8,26 +8,18 @@ part of 'selected_user.dart';
 
 SelectedUsers _$SelectedUsersFromJson(Map<String, dynamic> json) =>
     SelectedUsers(
-      totalCount: json['selectedTotalCount'] as int,
+      totalCount: (json['selectedTotalCount'] as num).toInt(),
       users: (json['users'] as List<dynamic>?)
           ?.map((e) => SelectedUser.fromJson(e as Map<String, dynamic>))
           .toList(),
     );
 
-Map<String, dynamic> _$SelectedUsersToJson(SelectedUsers instance) {
-  final val = <String, dynamic>{
-    'selectedTotalCount': instance.totalCount,
-  };
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('users', instance.users?.map((e) => e.toJson()).toList());
-  return val;
-}
+Map<String, dynamic> _$SelectedUsersToJson(SelectedUsers instance) =>
+    <String, dynamic>{
+      'selectedTotalCount': instance.totalCount,
+      if (instance.users?.map((e) => e.toJson()).toList() case final value?)
+        'users': value,
+    };
 
 SelectedUser _$SelectedUserFromJson(Map<String, dynamic> json) => SelectedUser(
       id: json['id'] as String?,
@@ -37,19 +29,12 @@ SelectedUser _$SelectedUserFromJson(Map<String, dynamic> json) => SelectedUser(
       favorite: json['favorite'] as bool?,
     );
 
-Map<String, dynamic> _$SelectedUserToJson(SelectedUser instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('id', instance.id);
-  val['uuid'] = instance.uuid;
-  writeNotNull('profile_nickname', instance.profileNickname);
-  writeNotNull('profile_thumbnail_image', instance.profileThumbnailImage);
-  writeNotNull('favorite', instance.favorite);
-  return val;
-}
+Map<String, dynamic> _$SelectedUserToJson(SelectedUser instance) =>
+    <String, dynamic>{
+      if (instance.id case final value?) 'id': value,
+      'uuid': instance.uuid,
+      if (instance.profileNickname case final value?) 'profile_nickname': value,
+      if (instance.profileThumbnailImage case final value?)
+        'profile_thumbnail_image': value,
+      if (instance.favorite case final value?) 'favorite': value,
+    };