+++ packages/kakao_flutter_sdk_talk/lib/src/model/friend.g.dart
@@ -15,20 +15,12 @@ Friend _$FriendFromJson(Map<String, dynamic> json) => Friend(
       json['allowed_msg'] as bool?,
     );
 
-Map<String, dynamic> _$FriendToJson(Friend instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('id', instance.id);
-  val['uuid'] = instance.uuid;
-  writeNotNull('profile_nickname', instance.profileNickname);
-  writeNotNull('profile_thumbnail_image', instance.profileThumbnailImage);
-  writeNotNull('favorite', instance.favorite);
-  writeNotNull('allowed_msg', instance.allowedMsg);
-  return val;
-}
+Map<String, dynamic> _$FriendToJson(Friend instance) => <String, dynamic>{
+      if (instance.id case final value?) 'id': value,
+      'uuid': instance.uuid,
+      if (instance.profileNickname case final value?) 'profile_nickname': value,
+      if (instance.profileThumbnailImage case final value?)
+        'profile_thumbnail_image': value,
+      if (instance.favorite case final value?) 'favorite': value,
+      if (instance.allowedMsg case final value?) 'allowed_msg': value,
+    };