+++ packages/kakao_flutter_sdk_template/lib/src/text_template.g.dart
@@ -16,20 +16,12 @@ TextTemplate _$TextTemplateFromJson(Map<String, dynamic> json) => TextTemplate(
       objectType: json['object_type'] as String? ?? "text",
     );
 
-Map<String, dynamic> _$TextTemplateToJson(TextTemplate instance) {
-  final val = <String, dynamic>{
-    'text': instance.text,
-    'link': instance.link.toJson(),
-  };
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('buttons', instance.buttons?.map((e) => e.toJson()).toList());
-  writeNotNull('button_title', instance.buttonTitle);
-  val['object_type'] = instance.objectType;
-  return val;
-}
+Map<String, dynamic> _$TextTemplateToJson(TextTemplate instance) =>
+    <String, dynamic>{
+      'text': instance.text,
+      'link': instance.link.toJson(),
+      if (instance.buttons?.map((e) => e.toJson()).toList() case final value?)
+        'buttons': value,
+      if (instance.buttonTitle case final value?) 'button_title': value,
+      'object_type': instance.objectType,
+    };